+++ checkpoint/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Reduce Memory Pressure from long held CPU buffers in Pathways MTC
 - Support explicit devices for Pathways MTC init
+- Avoid stale Pathways MTC worker dummy handles
 
 ## [0.12.1] - 2026-06-24